pyproject.toml
[project]
name = "agenticai"
version = "0.1.0"
description = "This is only a UV project created with an intent to learn and play around with Agentic AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.43.1",
    "ipython>=9.4.0",
    "openai>=1.100.2",
    "openai-agents>=0.2.9",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "py-hot-reload>=1.0.6",
]
